chore(docs-capture): share the fixture ids instead of copying them
From CodeRabbit on #9059. capture.ts carried its own copy of DOCS_IDS and the
admin credentials, kept in step by a comment. A renamed id would have sent a
shot to a 404 and still produced a plausible screenshot.

The ids move to docs-fixture-ids.ts, which has no imports and no side effects,
so capture.ts can share them without pulling in the seed script's PrismaClient
and production guard. seed-docs-fixtures.ts re-exports them for its own callers.
